Anvil Start
*UNTESTED CODE*
Started adding an anvil that follows the closest player to it, and falls on them if they stand under it for long enough.
Added a script to find and follow the closest player
Added a script to sense when players are near enough without using colliders or triggers (I don't know why, it would make things so much easier.). This uses events for when the sensor sucessfully senses a player, when the target changes, and conditions for whether it needs to sense for a certain amount of time first