Fading effect done!
I used an image to manipulate it's scale rect transform for the effect.
Please note that the update func needs to be removed from the actual game. It is there just to demonstrate the function's functionality. // To show what function will be called by other scripts to apply the fade effect.