Player Weapon Enabler/Disabler
Added a very quick way to enable/disable the player's weapon systems if they trigger a certain object. This uses a script I made in the previous commit for adding whitelist tags to colliders and triggers and calling "events" (programatically just functions) with the triggering/colliding object as well as keeping a list of triggering and colliding objects. To note: the system currently only has triggers implemented as they are what I needed at the time, future implementations will have the colliders and further "events" (mainly to do with the lists)