Section 2 Mostly Finished
Mostly finished integrating section 2. Only thing left is the pie bombs
Added a dumb way to do health pick ups that might need some more tweaking. It uses the existing TakeDamage function and detection in the player health system but now also checks if a triggering object is in the healable tags list, in which case we deal -1 damage to the player, ie. healing 1 damage