Lots of crucial changes.
Modularised the game initialiser further to support different prefab instantiation
Fixed the Audio transitions inside of the game scenes. It is also now very easy to set up and tweak.
Fixed the setting panel on escape pressing ! disabling the panel correctly if the panel was already enabled.
Added to the rhythm level.
Added tags to audio sources.
Made a shared prefab for music audio to play. etc